Split internal and API tests
