+++ src/test/java/one/util/streamex/AverageLongTest.java
